use variables instead of functions for spacing and typescales
the main reasons for functions were to throw warnings when a value is not on the given scale, but Sass will error anyway due to undefined variables. Secondly it was to make changing with vim increment/decrement shortcut easier but that's little reason to use them. Changing this back to variables makes autosuggest work nicely.